+++ src/app/modules/superAdmin/components/list-super-admin/list-super-admin.component.ts
@@ -25,6 +25,7 @@ export class ListSuperAdminComponent implements OnInit {
 
   showDeleteConfirm = false;
   adminIdToDelete: string | null = null;
+  deleteJustification = '';
 
   // Component-managed filter values
   filterValues: any = {
@@ -91,18 +92,20 @@ export class ListSuperAdminComponent implements OnInit {
   cancelDelete() {
     this.showDeleteConfirm = false;
     this.adminIdToDelete = null;
+    this.deleteJustification = '';
   }
 
   proceedDelete() {
-    if (this.adminIdToDelete) {
+    if (this.adminIdToDelete && this.deleteJustification.trim()) {
       this.onDelete(this.adminIdToDelete);
       this.showDeleteConfirm = false;
       this.adminIdToDelete = null;
+      this.deleteJustification = '';
     }
   }
 
   onDelete(adminId: string) {
-    this.superAdminService.deleteSuperAdmin(adminId).then((res: any) => {
+    this.superAdminService.deleteSuperAdmin(adminId, this.deleteJustification.trim()).then((res: any) => {
       if (this.globalService.handleSuccessService(res)) {
         // Refresh current view
         if (this.lastTableLazyLoadEvent) {
